requirements.txt
Mako==1.0.0
MarkupSafe==0.23
SQLAlchemy==0.9.8
alembic==0.7.1
amqp==1.4.6
anyjson==0.3.3
backports.ssl-match-hostname==3.4.0.2
beautifulsoup4==4.3.2
billiard==3.3.0.19
celery==3.1.17
certifi==14.05.14
flower==0.7.3
kombu==3.0.24
oauthlib==0.7.2
psycopg2==2.5.4
python-twitter==2.0
pytz==2014.10
redis==2.10.3
requests==2.5.0
requests-oauthlib==0.4.2
simplejson==3.6.5
tornado==4.0.2
